Split provider into read and write interfaces
