--- lib/src/utils/index.dart
@@ -1,9 +1,0 @@
-import 'dart:convert';
-
-export 'convert.dart';
-export 'rpc.dart';
-
-prettyPrintJson(Map<String, dynamic> json) {
-  var encoder = JsonEncoder.withIndent("  ");
-  print(encoder.convert(json));
-}
