+++ lib/starknet.dart
@@ -3,10 +3,9 @@
 /// It aims at supporting the same API as the official Starknet API, but in Dart.
 library starknet;
 
+export 'src/crypto/index.dart';
 export 'src/provider/index.dart';
 export 'src/signer.dart';
-export 'src/account.dart';
 export 'src/static_config.dart';
-export 'src/field_element.dart';
-export 'src/crypto/index.dart';
-export 'src/utils/index.dart';
+export 'src/felt.dart';
+export 'src/convert.dart';
