+++ test/utils.dart
@@ -1,6 +1,17 @@
 import 'package:starknet/starknet.dart';
 
-Provider getJsonRpcProvider() {
+ReadProvider getJsonRpcProvider() {
+  const network = String.fromEnvironment('NETWORK', defaultValue: 'testnet');
+  if (network == 'testnet') {
+    return JsonRpcReadProvider.infuraGoerliTestnet;
+  } else if (network == 'mainnet') {
+    return JsonRpcReadProvider.infuraMainnet;
+  } else {
+    return JsonRpcReadProvider.devnet;
+  }
+}
+
+Provider getJsonRpcAccount() {
   const network = String.fromEnvironment('NETWORK', defaultValue: 'testnet');
   if (network == 'testnet') {
     return JsonRpcProvider.infuraGoerliTestnet;
@@ -8,16 +19,5 @@ Provider getJsonRpcProvider() {
     return JsonRpcProvider.infuraMainnet;
   } else {
     return JsonRpcProvider.devnet;
-  }
-}
-
-Account getJsonRpcAccount() {
-  const network = String.fromEnvironment('NETWORK', defaultValue: 'testnet');
-  if (network == 'testnet') {
-    return JsonRpcAccount.infuraGoerliTestnet;
-  } else if (network == 'mainnet') {
-    return JsonRpcAccount.infuraMainnet;
-  } else {
-    return JsonRpcAccount.devnet;
   }
 }
